Row lost while extracting Polygon data from CSV
This row contains the data that had to be removed in order for excell to read it. I am not sure what it is. But it does not contain KID data or any other form of data. I am assuming its a polygon for the whole of kelowna. I feel it might be ok to leave it out of the code